# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Job Finder API",
        "description": "API para el sistema de búsqueda y gestión de empleos",
        "contact": {
            "name": "Soporte Job Finder",
            "email": "admin@jobfinder.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "Servidor de desarrollo"
        }
    ],
    "paths": {
        "/api/applications/my": {
            "get": {
                "tags": [
                    "Postulaciones"
                ],
                "summary": "Obtener mis postulaciones",
                "description": "Obtiene las postulaciones del usuario autenticado",
                "operationId": "getMyApplications",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Número de elementos por página",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 10
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filtrar por estado",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "pendiente",
                                "revisando",
                                "aceptada",
                                "rechazada"
                            ],
                            "example": "pendiente"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Mis postulaciones obtenidas exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Mis postulaciones obtenidas exitosamente"
                                        },
                                        "data": {
                                            "properties": {
                                                "data": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "object"
                                                    }
                                                },
                                                "current_page": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "per_page": {
                                                    "type": "integer",
                                                    "example": 10
                                                },
                                                "total": {
                                                    "type": "integer",
                                                    "example": 5
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/applications": {
            "post": {
                "tags": [
                    "Postulaciones"
                ],
                "summary": "Crear nueva postulación",
                "description": "Crea una nueva postulación para un trabajo",
                "operationId": "createApplication",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "trabajo_id",
                                    "cv"
                                ],
                                "properties": {
                                    "trabajo_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "cv": {
                                        "description": "Archivo CV en PDF",
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "mensaje": {
                                        "type": "string",
                                        "example": "Me interesa mucho esta oportunidad..."
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Postulación enviada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Postulación enviada exitosamente"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Error de validación"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/applications/{id}/status": {
            "put": {
                "tags": [
                    "Administración"
                ],
                "summary": "Actualizar estado de postulación",
                "description": "Actualiza el estado de una postulación (solo administradores)",
                "operationId": "updateApplicationStatus",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la postulación",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "estado"
                                ],
                                "properties": {
                                    "estado": {
                                        "type": "string",
                                        "enum": [
                                            "pendiente",
                                            "revisando",
                                            "aceptada",
                                            "rechazada"
                                        ],
                                        "example": "revisando"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Estado de postulación actualizado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Estado de postulación actualizado correctamente"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Postulación no encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Postulación no encontrada"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/jobs/{jobId}/applications": {
            "get": {
                "tags": [
                    "Administración"
                ],
                "summary": "Obtener postulaciones de un trabajo",
                "description": "Obtiene todas las postulaciones de un trabajo específico",
                "operationId": "getJobApplications",
                "parameters": [
                    {
                        "name": "jobId",
                        "in": "path",
                        "description": "ID del trabajo",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Número de elementos por página",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 10
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filtrar por estado",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "pendiente",
                                "revisando",
                                "aceptada",
                                "rechazada"
                            ],
                            "example": "pendiente"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Postulaciones del trabajo obtenidas exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Postulaciones del trabajo obtenidas exitosamente"
                                        },
                                        "data": {
                                            "properties": {
                                                "data": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "object"
                                                    }
                                                },
                                                "current_page": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "per_page": {
                                                    "type": "integer",
                                                    "example": 10
                                                },
                                                "total": {
                                                    "type": "integer",
                                                    "example": 25
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/admin/applications/stats": {
            "get": {
                "tags": [
                    "Administración"
                ],
                "summary": "Obtener estadísticas de postulaciones",
                "description": "Obtiene estadísticas detalladas de las postulaciones",
                "operationId": "getApplicationStats",
                "responses": {
                    "200": {
                        "description": "Estadísticas de postulaciones obtenidas exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Estadísticas de postulaciones obtenidas exitosamente"
                                        },
                                        "data": {
                                            "properties": {
                                                "total_applications": {
                                                    "type": "integer",
                                                    "example": 500
                                                },
                                                "applications_by_status": {
                                                    "type": "object"
                                                },
                                                "applications_by_month": {
                                                    "type": "object"
                                                },
                                                "top_applicants": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}